[FW] Added "CANReceiver_State_TypeDef state" field to CANReceiver_TypeDef. Added checking GPSFixType and according led DEBUG1 blinking to ActionScheduler . Added FIFOQueue_clear method. Added _GPSDriver_parseGPSFixType method to GPSDriver. Added fixType field to Ublox8MGPSDriver_TypeDef. Added GPSDriver_getFixType() method to Ublox8MGPSDriver. Added _GPSDriver_FixTypeRawUARTCharHandler() raw uart byte received callback. Changed frequency of sending frames in CANBusMock SW.

+++ FW/Inc/user/string_operations.h
@@ -30,7 +30,7 @@ StringOperations_Status_TypeDef	stringToFixedPoint(uint8_t* sentence, uint16_t l
 StringOperations_Status_TypeDef	decChar2Uint8(uint8_t c, uint8_t* pRetInt);
 StringOperations_Status_TypeDef	hexChar2Uint8(uint8_t c, uint8_t* pRetInt);
 
-StringOperations_Status_TypeDef	string2UInt32(uint8_t* sentence, uint8_t length, uint32_t* pRetInt);
+StringOperations_Status_TypeDef	string2UInt32(uint8_t* sentence, uint16_t length, uint32_t* pRetInt);
 
 StringOperations_Status_TypeDef	uInt8ToString(uint8_t* pRetBuffer, uint8_t val);
 StringOperations_Status_TypeDef appendUInt8ToString(uint8_t* pRetBuffer, uint8_t val, uint16_t bufferSize);
